NEEDS FIXUP: BUG-2625: Fix HashMap ordering assumptions
This is a preliminary patch to mark validations which are failing with
JDK8. These fail because they make explicit assumptions about how
HashMap orders its entries.

I am not sure whether the test are bad or H2 plugin fails to
canonicalize map entry ordering (which is not guaranteed by API contract
and change at any time).

Change-Id: I7c4c302f4dd7b1b5ad3c54a7be452424c53f8e08
Signed-off-by: Robert Varga <robert.varga@pantheon.sk>