Temporarily remove h2 data store feature
There was an issue reported by Thanh and Phil that TSDR is pulling
msql-connector-java component into ODL distribution. Since
msql-connector-java carries GPL license, this would have impact on
ODL distribution.

We believe the root cause is the pax jdbc repository that H2 data
store is using when installing the necessary features needed by H2.
In order to verify this, we decided to temporarily disable H2 and
remove the pax jdbc dependency from TSDR. At the same time we added
hsqldb to replace H2 until further investigation and solution could
be found regarding this issue in h2.

This patch set does the following:
  1. Temporarily removed h2 data store from pom.xml and features.xml.
  2. To make sure that we pass integration test where odl-tsdr-h2-all
was added in integration test, we temporarily point this feature to
 hsqldb.

We're planning to merge this patch set and ask Thanh to verify from
ODL distribution that if the msql-connector-java component is gone.

Change-Id: I8db1af8bd4ec132c21b8bebc8cf2f12243a052cf
Signed-off-by: YuLing Chen <yuling_c@dell.com>