Simplify threading model in openflow collector
TSDRInventoryNodesPoller and TSDROpenflowCollector use dedicated Threads
that sleep and communicate via wait/notify. Also wait/notify is used to
shutdown the threads. This can be simplified by using a
single-threaded ScheduledExecutor in TSDRInventoryNodesPoller and a
single-threaded Executor in TSDROpenflowCollector for storing the records.
The scheduled task TSDRInventoryNodesPoller just notifies the
TSDROpenflowCollector to simply submit a task to its Executor.

Also, TSDROpenflowCollector stored the TSDRMetricRecordBuilderContainers
in an array with the indexes stored in the id2Index Map. However this is
essentially equivalent to a Map so it was converted to a HashMap which
is simpler than synchronizing 2 structures and growing/shrinking the array.

With these changes, the TSDRInventoryNodesPoller was substanially reduced
and called into TSDROpenflowCollector to do most of the work so it no longer
seemed useful and thus was removed.

A unit test was added to test end-to-end polling. To facilitate setting the
poll interval low for the tests, the TSDROSCConfig is now passed to
the TSDROpenflowCollector - previously it wrote a default config in the
init method which overwrites use changes on restarts. The BP XML uses
the <clustered-app-config> element to inject the TSDROSCConfig.

Change-Id: Icc6e922bc7e23fe6539134774ef736683ca8742d
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>