Remove external file logging in TSDROpenflowCollector
Apparently this was put in so the original author could separate
TSDROpenflowCollector logging from the karaf.log. However this can
be done by using the sift appender in the pax logging cfg. Also
there's no precendence for directing a specific class's log
output to a separate file elsewhere in ODL and I dont think it's
a practice we should employ.

Change-Id: Ie94a24302e5eab8c6bcca5037292701f0c66a939
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>