Simplify threading model in snmp collector
SNMPDataCollector and SNMPInterfacePoller use dedicated Threads
that sleep and communicate via wait/notify. Also wait/notify is
used to shutdown the threads. This can be simplified by using a
single-threaded ScheduledExecutor for the polling task and a
single-threaded Executor for storing the records. Also, the code
processes each IP in turn, ie obtains the interface data then stores
the records, and it really doesn't need to store the intermediate
data in the SNMPDataCollector - it's further simplified by
passing the intermediate data locally.

The getInterfaces RPC was previously invoked in a blocking manner -
is now handled async via a Future callback which submits a task to
the storing executor.

With these changes, the SNMPInterfacePoller was substanially reduced
and called into SNMPDataCollector to do most of the work so it no longer
seemed useful and thus was removed.

Change-Id: Ic101bbbd2783094eba7ea7381f64b726c7677268
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>