Fix create table errors in hbase
HBaseConfiguration#create throws an exception complaining that the
supplied hbase.defaults.for.version appeaars to be older than the
library's version 0.94.15. However the hbase-default.xml in the bundle
does specify version 0.94.15 so that and the fact that the message
prints "null" for the found version indicates it didn't find our
hbase-default.xml file.

Looking at the code it attempts to load it from the Thread context
ClassLoader, if set, or from its own ClassLoader. Either way it's not
our bundle's ClassLoader so it doesn't find our file. I modified
HBaseDataStore to call Thread.currentThread().setContextClassLoader(...)
approriately.

Also the hbase-default.xml (and hbase-site.xml) are under configuration/initial
but should be at the top-level of the bundle.

Change-Id: I37dd9d4413c5cf7bcd6cb9a308d2396ce1c79a1b
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>