Remove static SchedulerService instance
Make it a proper singleton with an interface and implementation
advertised as an OSGi service. SchedulerService was changed to
an interface. The SchedulerServiceImpl was put in a new bundle,
tsdr-persistence-impl, to separate the impl from the interface
and to avoid the BP xml from being scanned by other bundles that
use the maven-blueprint-bundle.

The SchedulerService methods were also changed to return
ListenableScheduledFuture so users can wait async for completion.

Change-Id: I7dea6516e117ae3e64ec240c5c8474c702775530
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>