Refactor TsdrRHBasePersistenceServiceImpl and CreateTableTask
Refactored the relationship between these 2 classes and added
retries for data store operations on occurrence of a
TableNotFoundException. Previously if a TableNotFoundException
occured, a new CreateTableTask (if one was not already running)
was created but the current operation just silently exited
without retrying the operation when the CreateTableTask completed.

Also data store operations now wait for the current CreateTableTask
to complete.

The CreateTableTask now reschedules itself each time if necessary
after execution instead of scheduling at a fixed rate and cancelling
itself. This avoids needing the ScheduledFuture injected which is
problematic.

Beefed up UTs...

Change-Id: I881e849ec578843c2bb6d6e5ab0ab464653dc0bb
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>