Remove Task class
The Task basically wraps Runnable and adds a setScheduledFuture method.
setScheduledFuture is called by the SchedulerService after the task is
scheduled however this is problematic as it's possible for the scheduled
task to run before setScheduledFuture is called (especially if initial delay
is 0), which could lead to unexpected behavior. There are no longer any
users of the ScheduledFuture so we can eliminate it from the API which then
renders the Task class unnecessary so remove it altogther.

Change-Id: I879c6430a1563886c8487db64d4b4794c8268afe
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>