Simplify ControllerMetricCollector
The CpuDataCollector interface returns Optional data so we really
don't need the CpuDataCollector itself to be optional in the
ControllerMetricCollector. If there's no real implementation of
CpuDataCollector available then use a null instance that returns
empty Optionals.

Also use OptionalDouble instead of Optional<Double> for efficiency.

Also streamlined the code in ControllerMetricCollector.

Change-Id: I7284cc1959e73781e9971dcc0dcc228d043b28ac
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>