Simplify SyslogFilter
The interface was simplified to just the matches(...) API. The
getPersistenceDestination API was not used and not clearly defined,
specifically as to what the return int value signifies. For the
filterAndParseSyslog API, it seems the "filter" aspect should already
be handled by matches - it also seems that building the TSDRLogRecord
shouldn't be related to the filtering and should be consistent. There is
only one implementation of SyslogFilter that basically is a no-op so it's
hard to know what the original intention was.

Also, the map in the SyslogFilterManager was changed from SyslogFilter[] to
SyslogFilter. It's unclear why this was an array - the code only expected one
element and, even if it allowed multiple, the applyFilters method only returns
a single TSDRLogRecord. So again, unclear as to what the original intention was.

Change-Id: I8264056f0bb48d3edbbf6b918b0dcfbccffd6107
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>