Improve NetflowPacketParser
Separated out v5 and v9 NetflowPacketParser classes
for clarity with NetflowPacketParser as the abstract base.

Instead of using hard-coded offsets into the byte stream
which is error prone, the current position is tracked and
updated automatically.

The v5 code was mostly correct except the sampling_interval field
is 14 bits and not 6.

The v9 code was incorrect for several reasons wrt
https://www.ietf.org/rfc/rfc3954.txt, in particular:

 - It assumed there is only one template when parsing a data flow set.

 - If there was no template, it assumed a particular template format.
   A data flow set cannot be parsed without a corresponding template.

 - When parsing a template, the template ID and field count were not
   parsed and thus were interpreted as a field record.

The code also assumes a template flow set is always in the same packet
as the data flow sets but templates are mostly sent once and are supposed
to be cached in the collector to be used to decode subsequent data flow
set packets. This will be addressed in another patch.

Change-Id: I74c0f3bdae266418b34195e25aa35167fb43a5df
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>